improves response handling
This commit updates the response handling to no longer assume the presence of a "content" value on the basis of "message" and "choices" being in the response. If a 200 OK response is received, but the structure of the response doesn't match our expectations, we will propagate back to Leo an InvalidResponseFormat error.